use ObjectID scalar instead of ID for _id field in gql schema

+++ bin/clean_db.js
@@ -2,12 +2,18 @@
 export async function cleanDB(db) {
     const collections = await db.collections()
     console.log(`Found ${collections.length} collections`)
+
+    if (!collections.length) return
+
     const cleanResult = await Promise.all(collections.map((c) => c.drop()))
+
     const allCleaned = cleanResult.every((x) => x)
+
     if (allCleaned) {
         console.log("All collections removed")
     } else {
         throw new Error("Not all collections could be removed.")
     }
+
     return
 }
